+++ src/core/lib/CHANGELOG.md
@@ -6,7 +6,7 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and 
 
 ## [Unreleased]
 
-## [0.2.0] - 2026-06-10
+## [0.2.0] - 2026-06-16
 
 ### Added